requirements.txt
anyio==4.9.0
asgiref==3.8.1
asyncio==3.4.3
black==25.1.0
click==8.2.1
Django==5.2.3
Faker==37.4.0
graphql-core==3.2.6
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.0
strawberry-graphql==0.273.3
strawberry-graphql-django==0.62.0
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
